% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{adaptive_imputation}
\alias{adaptive_imputation}
\title{Create an Adaptive Imputation object}
\usage{
adaptive_imputation(u, d, v, alpha, ...)
}
\arguments{
\item{u}{A \emph{matrix} "left singular-ish" vectors.}

\item{d}{A \emph{vector} of "singular-ish" values.}

\item{v}{A \emph{matrix} of "right singular-ish" vectors.}

\item{alpha}{Value of \code{alpha} after final iteration.}

\item{...}{Optional additional items to pass to the constructor.}
}
\value{
An \code{adaptive_imputation} object.
}
\description{
\code{adaptive_imputation} objects are a subclass of
\code{\link[LRMF3:svd_like]{LRMF3::svd_like()}}, with an additional field \code{alpha}.
}
