% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-impute.R
\name{adaptive_impute}
\alias{adaptive_impute}
\title{AdaptiveImpute}
\usage{
adaptive_impute(
  X,
  rank,
  initialization = c("svd", "adaptive-initialize"),
  epsilon = 1e-07,
  max_iter = 200L,
  check_in = 1L,
  verbose = FALSE
)
}
\arguments{
\item{X}{A sparse matrix of \code{sparseMatrix} class.}

\item{rank}{Desired rank (integer) to use in the low rank approximation.}

\item{initialization}{TODO.}

\item{epsilon}{Convergence criteria, measured in terms of relative change
in Frobenius norm of the full imputed matrix. Defaults to \code{1e-7}.}

\item{max_iter}{Maximum number of iterations to perform (integer). Defaults
to \code{200L}. In practice 10 or so iterations will get you a decent
approximation to use in exploratory analysis, and and 50-100 will get
you most of the way to convergence.}

\item{verbose}{TODO.}
}
\value{
A \code{low_rank_matrix_factorization} object.
}
\description{
An implementation of the \code{AdaptiveImpute} algorithm for matrix completion
for sparse matrices.
}
\examples{

mf <- adaptive_impute(ml100k, rank = 5L, max_iter = 10L, verbose = TRUE)

mf

# reconstruct a rank-5 approximation of M
s$u \%*\% diag(s$d) \%*\% t(s$v)

# build a rank-5 approximation to M only for
# observed elements of M
masked_approximation(s, M)

}
