% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-initialize.R
\name{adaptive_initialize}
\alias{adaptive_initialize}
\title{AdaptiveInitialize}
\usage{
adaptive_initialize(X, rank, p_hat = NULL, ...)
}
\arguments{
\item{X}{A sparse matrix of \code{sparseMatrix} class. Explicit (observed)
zeroes in \code{X} can be dropped for}

\item{rank}{Desired rank (integer) to use in the low rank approximation.}

\item{p_hat}{The portion of \code{X} that is observed. Defaults to \code{NULL},
in which case \code{p_hat} is set to the number of observed elements of
\code{X}. Useful when you have many explicit zeros that w}

\item{...}{Ignored.}
}
\value{
A low rank matrix factorization represented by an
\code{LRMF} object. See \code{\link[LRMF3:svd_like]{LRMF3::svd_like()}} for details.
}
\description{
An implementation of the \code{AdaptiveInitialize} algorithm for
matrix imputation for sparse matrices. At the moment the implementation
is only suitable for small matrices with on the order of thousands
of rows and columns at most.
}
\examples{

mf <- adaptive_initialize(ml100k, rank = 3L)
mf

# build a rank-5 approximation only for
# observed elements of ml100k
preds <- predict(mf, ml100k)

}
