% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper-triangular.R
\name{citation_adaptive_impute}
\alias{citation_adaptive_impute}
\title{AdaptiveImpute (sparse)}
\usage{
citation_adaptive_impute(
  M,
  r,
  epsilon = 1e-07,
  max_iter = 200,
  additional = 100,
  check_in = 50
)
}
\arguments{
\item{M}{A sparse Matrix created by one of the \link{Matrix} pkg
constructors.}

\item{r}{Desired rank to use in the low rank approximation.}

\item{epsilon}{Tolerance, measured in terms of relative change
in Frobenius norm of the full imputed matrix.}
}
\value{
A list with elements:
\itemize{
\item \code{u}: Left singular-ish vectors
\item \code{d}: Singular-ish values
\item \code{v}: Right singular-ish vectors
}
}
\description{
A reference implemention of the \code{AdaptiveImpute} algorithm using sparse
matrix computations. For citation matrices where missing values in
the upper triangle are taken to be \emph{explicitly observed} observed
zeros, as opposed to missing values.
}
\examples{

library(Matrix)

set.seed(27)

M <- rsparsematrix(12, 12, nnz = 30)

s <- citation_adaptive_impute(M, 5, max_iter = 20)
s

# reconstruct a rank-5 approximation of M
s$u \%*\% diag(s$d) \%*\% t(s$v)

# build a rank-5 approximation to M only for
# observed elements of M
masked_approximation(s, M)

}
