% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-adi.R
\name{eigInit}
\alias{eigInit}
\title{Title}
\usage{
eigInit(M, r)
}
\arguments{
\item{M}{sparse matrix}

\item{r}{imputation rank}
}
\description{
eigInit and its subfunction Mx are to make the svd in the initalizer.
}
\details{
given sparse M compute eigen of
t(M)%*%M/p^2 - (1-p)\emph{diag(t(M)%}%M)
where p is the mean nnz of M.
}
