% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastadi-package.R
\docType{package}
\name{fastadi-package}
\alias{fastadi}
\alias{fastadi-package}
\title{fastadi: Self-Tuning Data Adaptive Matrix Imputation}
\description{
Implements the AdaptiveImpute matrix completion algorithm of 'Intelligent Initialization and Adaptive Thresholding for Iterative Matrix Completion' \doi{10.1080/10618600.2018.1518238} as well as the specialized variant of 'Co-Factor Analysis of Citation Networks' \doi{10.1080/10618600.2024.2394464}. AdaptiveImpute is useful for embedding sparsely observed matrices, often out performs competing matrix completion algorithms, and self-tunes its hyperparameter, making usage easy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rohelab.github.io/fastadi/}
  \item \url{https://github.com/RoheLab/fastadi}
  \item Report bugs at \url{https://github.com/RoheLab/fastadi/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Juhee Cho
  \item Donggyu Kim
  \item Karl Rohe
}

}
\keyword{internal}
