% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{masked_approximation}
\alias{masked_approximation}
\title{Expand an SVD only at observed values of a sparse matrix}
\usage{
masked_approximation(s, mask)
}
\arguments{
\item{s}{An svd-like list with elements \code{u}, \code{d}, \code{v}, where \code{u} and \code{v}
are low-rank matrices and \code{d} is a vector.}

\item{mask}{A sparse matrix that can be coerced into triplet form.}
}
\value{
A reconstruction of \code{mask}A \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} of the same
}
\description{
TODO: describe what it looks like for dimensions to match up between
\code{s} and \code{mask}. See \code{vignette("sparse-computations")} for mathematical
details.
}
\examples{

# U \%*\% diag(d) \%*\% t(V)

}
