% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-adi.R
\name{Mx}
\alias{Mx}
\title{Title}
\usage{
Mx(x, args)
}
\arguments{
\item{x}{TODO}

\item{args}{TODO}
}
\description{
Has the effect of  \link[= t(M)\%*\%M - (1-p)*diag(t(M)\%*\%M) ]{ t(M)\%*\%M - (1-p)diag(t(M)\%\%M) } *x
}
