% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-initialize.R
\name{sparse_adaptive_initialize}
\alias{sparse_adaptive_initialize}
\title{AdaptiveInitialize (sparse)}
\usage{
sparse_adaptive_initialize(M, r, additional = min(100, ncol(M)/4))
}
\arguments{
\item{M}{A sparse Matrix created by one of the \link{Matrix} pkg
constructors.}

\item{r}{Desired rank to use in the low rank approximation.}
}
\value{
A list with elements:
\itemize{
\item \code{u}: Left-singular-ish vectors
\item \code{d}: Singular-ish values
\item \code{v}: Right-singular-ish vectors
}
}
\description{
A reference implemention of the \code{AdaptiveInitialize} using sparse
matrix computations.
}
\examples{

library(Matrix)

set.seed(27)

# create a random 8 x 12 sparse matrix with 30 nonzero entries
M <- rsparsematrix(8, 12, nnz = 30)

s <- sparse_adaptive_initialize(M, 5)
s

# reconstruct a rank-5 approximation of M
s$u \%*\% diag(s$d) \%*\% t(s$v)

# build a rank-5 approximation to M only for
# observed elements of M
masked_approximation(s, M)

}
