% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastadi-package.R
\docType{package}
\name{fastadi-package}
\alias{fastadi}
\alias{fastadi-package}
\title{fastadi: Data Adaptive Matrix Imputation}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/fastadi}
  \item Report bugs at \url{https://github.com/alexpghayes/fastadi/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

Authors:
\itemize{
  \item Juhee Cho
  \item Donggyu Kim
  \item Karl Rohe
}

}
\keyword{internal}
